web/requirements.txt
Flask==2.1.1
google-cloud-firestore==2.4.0
google-cloud-pubsub==2.12.1
requests==2.27.1
gunicorn==20.1.0
